requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cramjam==2.9.0
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.15
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.6
dvc==3.57.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dvclive==3.48.0
entrypoints==0.4
executing==2.1.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.16.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl-lock==8.1.0
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
funcy==2.0
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
grandalf==0.8
grpcio==1.68.0
gto==1.7.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
iterative-telemetry==0.0.9
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
live==0.0.3
llvmlite==0.43.0
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlip==0.2
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook-shim==0.2.4
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
opt-einsum==3.4.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
polars==1.16.0
prometheus-client==0.21.1
prompt-toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.0.0
pybaseball==2.2.7
pycparser==2.22
pydantic==2.10.2
pydantic-core==2.27.1
pydot==3.0.2
pygit2==1.16.0
pygithub==2.5.0
pygments==2.18.0
pygtrie==2.5.0
pyjwt==2.10.1
pynacl==1.5.0
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.12
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.9
seaborn==0.13.2
semver==3.0.2
send2trash==1.8.3
setuptools==75.6.0
shap==0.46.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqltrie==0.11.1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.14.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
yarl==1.18.0
zc-lockfile==3.0.post1
